Move the in-flight counter to its own module
Verbatim move out of the session registry, which is about to be deleted
while the counter stays.
