+++ src/transport/sessionRegistry.ts
@@ -1,4 +1,3 @@
-import type { RequestHandler } from 'express';
 import type { NodeStreamableHTTPServerTransport } from '@modelcontextprotocol/node';
 import { logger } from '../runtime/logger.js';
 
@@ -9,23 +8,6 @@ export type SessionEntry = {
 };
 
 export type SessionRegistry = { [sessionId: string]: SessionEntry };
-
-export interface InFlightCounter {
-	readonly middleware: RequestHandler;
-	readonly count: () => number;
-}
-
-export function createInFlightCounter(): InFlightCounter {
-	let n = 0;
-	const middleware: RequestHandler = (_req, res, next) => {
-		n++;
-		res.on('close', () => {
-			n--;
-		});
-		next();
-	};
-	return { middleware, count: () => n };
-}
 
 // Marks a session as having an in-flight request or open response stream:
 // increments the active-request count and cancels any pending idle expiry.
